+++ packages/appcd-config/test/fixtures/bad-es-export.js
@@ -1,5 +1,0 @@
-const cfg = {
-	name: 'foo'
-};
-
-export default cfg;