+++ packages/appcd-config/test/fixtures/good-load.js
@@ -1,4 +1,0 @@
-module.exports = {
-	name: 'foo',
-	age: 21
-};