+++ packages/appcd-config/test/fixtures/good-readonly.js
@@ -1,8 +1,0 @@
-module.exports = {
-	name: 'foo',
-
-	/**
-	 * @readonly
-	 */
-	title: 'bar'
-};