+++ packages/appcd-core/src/status-monitor.js
@@ -1,5 +1,4 @@
 import Agent from 'appcd-agent';
-import Config from 'appcd-config';
 import gawk from 'gawk';
 import os from 'os';
 import Response, { codes } from 'appcd-response';
@@ -26,10 +25,6 @@ export default class StatusMonitor extends DataServiceDispatcher {
 	 * @access public
 	 */
 	constructor(cfg) {
-		if (!(cfg instanceof Config)) {
-			throw new TypeError('Expected config to be a Config object');
-		}
-
 		super();
 
 		/**